#3 Renaming to fragment message
