+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,9 +7,9 @@
 
     <fragment
         android:id="@+id/FirstFragment"
-        android:name="com.josdem.authenticator.FirstFragment"
+        android:name="com.josdem.authenticator.MessageFragment"
         android:label="@string/first_fragment_label"
-        tools:layout="@layout/fragment_first">
+        tools:layout="@layout/fragment_message">
 
         <action
             android:id="@+id/action_FirstFragment_to_SecondFragment"
